ported sketch with merged sketch and bootloader from legacy into integration test
