+++ internal/integrationtest/compile_4/testdata/SketchWithMergedSketchAndBootloader/SketchWithMergedSketchAndBootloader.ino
@@ -1,0 +1,7 @@
+void setup() {
+
+}
+
+void loop() {
+
+}
