+++ internal/integrationtest/compile_4/testdata/SketchWithMergedSketchAndBootloader/header.h
@@ -1,0 +1,1 @@
+#define TRUE FALSE
