+++ internal/integrationtest/compile_4/testdata/SketchWithMergedSketchAndBootloader/other.ino
@@ -1,0 +1,3 @@
+String hello() {
+  return "world";
+}
