--- legacy/builder/test/sketch1/header.h
@@ -1,1 +1,0 @@
-#define TRUE FALSE
