--- legacy/builder/test/sketch1/merged_sketch.txt
@@ -1,15 +1,0 @@
-#include <Arduino.h>
-#line 1 {{QuoteCppString .sketch.MainFile}}
-void setup() {
-
-}
-
-void loop() {
-
-}
-#line 1 {{QuoteCppString (index .sketch.OtherSketchFiles 0).Name}}
-
-#line 1 {{QuoteCppString (index .sketch.OtherSketchFiles 1).Name}}
-String hello() {
-  return "world";
-}
