--- legacy/builder/test/sketch1/other.ino
@@ -1,3 +1,0 @@
-String hello() {
-  return "world";
-}
