--- legacy/builder/test/sketch1/sketch1.ino
@@ -1,7 +1,0 @@
-void setup() {
-
-}
-
-void loop() {
-
-}
